Convert Resizer into a hook/component
