+++ src/components/App.js
@@ -6,6 +6,7 @@ import Network from './tabs/network/Network';
 import { useState } from 'preact/hooks';
 import useConsoleData from './useConsoleData';
 import useNetworkData from './useNetworkData';
+import useResizer from './useResizer';
 
 const defaultHeight = 350;
 
@@ -15,40 +16,18 @@ const defaultHeight = 350;
 export default () => {
   const ref = createRef();
   const [isOpen, setOpen] = useState(false);
-  const [height, setHeight] = useState(defaultHeight);
   const [activeTab, setActiveTab] = useState('console');
   const consoleData = useConsoleData();
   const networkData = useNetworkData();
-
-  const tabContents = {
-    console: <Console consoleData={consoleData}/>,
-    network: <Network networkData={networkData}/>
-  };
-
   const toggleOpen = () => {
     ref.current && (ref.current.style.height = isOpen ? '0px' : `${height}px`);
     setOpen(!isOpen);
   };
+  const [resizer, height] = useResizer({ close: toggleOpen, defaultHeight, targetRef: ref });
 
-  const resizeMouseDown = (event) => {
-    window.addEventListener('mousemove', resizeMouseMove, false);
-    window.addEventListener('mouseup', resizeMouseUp, false);
-  };
-
-  const resizeMouseMove = (event) => {
-    const height = Math.min(window.innerHeight, window.innerHeight - event.clientY);
-    ref.current && (ref.current.style.height = `${height}px`);
-    if (height < 20) {
-      toggleOpen();
-      resizeMouseUp();
-      setHeight(defaultHeight);
-    }
-  };
-
-  const resizeMouseUp = (event) => {
-    window.removeEventListener('mousemove', resizeMouseMove, false);
-    window.removeEventListener('mouseup', resizeMouseUp, false);
-    setHeight(parseFloat(ref.current.style.height) || defaultHeight);
+  const tabContents = {
+    console: <Console consoleData={consoleData}/>,
+    network: <Network networkData={networkData}/>
   };
 
   const onClickTab = (event) => setActiveTab(event.target.textContent.toLowerCase());
@@ -60,7 +39,7 @@ export default () => {
     <div className={`${styles.app} ${isOpen ? styles.open : ''}`} ref={ref}>
       {!isOpen ? icon : (
         <Fragment>
-          <div className={styles.resizer} onMouseDown={resizeMouseDown}/>
+          {resizer}
           {icon}
           <div className={styles.tabBar}>
             <div className={getTabClassName('console')} onClick={onClickTab}>Console</div>
