+++ src/components/dataTree/Node.module.scss
@@ -89,7 +89,7 @@
   }
 
   .caretIcon {
-    width: 8px;
+    width: 6px;
     text-align: right;
   }
